Ensure loop values are sent on 'looping' change
[live.observer] only outputs when the value of the property changes.
This leads to difficulties when changing 'looping' from 0 to 1 because
i.e. 'loop_start' doesn't output, when clip start AND loop start are
'1.1.1'. Usage of [pack] and [buddy] is broken.

Solution:

Bang the observers for 'loop_start' and 'loop_end' when 'looping'
changes.

Observe, whether this behaviour leads to problems with other properties.